Add files via upload
Current state of the map.  Previous versions were in a perforce repository on a local hard drive.... somewhere.  Next up... I want to make it load faster and I want to get rid of some features that didn't pan out (VCR-style controls and scanning the server for the file names instead of relying on a manifest file)